+++ Evergreen/Apps/Get-Anaconda.ps1
@@ -1,4 +1,4 @@
-Function Get-Anaconda {
+function Get-Anaconda {
     <#
         .SYNOPSIS
             Get the current version and download URL for Anaconda.
@@ -16,50 +16,44 @@ Function Get-Anaconda {
         $res = (Get-FunctionResource -AppName ("$($MyInvocation.MyCommand)".Split("-"))[1])
     )
 
-    # Construct the Anaconda repo uri
-    $Uri = $res.Get.Update.Uri -replace "#replace", $res.Get.Update.ReplaceFileList
     # Query the repo to get the full list of files
-    $updateFeed = Invoke-RestMethodWrapper -Uri $Uri
-
-    If ($Null -ne $updateFeed) {
+    $Uri = $res.Get.Update.Uri -replace "#replace", $res.Get.Update.ReplaceFileList
+    $UpdateFeed = Invoke-RestMethodWrapper -Uri $Uri
+    if ($null -ne $UpdateFeed) {
 
         # Grab the Windows files
-        $FileNames = $updateFeed.PSObject.Properties.name -match $res.Get.MatchFileTypes
+        $FileNames = $UpdateFeed.PSObject.Properties.name -match $res.Get.MatchFileTypes
 
-        # Grab all the version numbers
-        Try {
-            $AllVersions = [RegEx]::Matches($FileNames, $res.Get.MatchVersion) | Select-Object -ExpandProperty Value -Unique
+        try {
+            # Grab all the version numbers; Grab latest version number
+            $AllVersions = [RegEx]::Matches($FileNames, $res.Get.MatchVersion) | Select-Object -ExpandProperty "Value" -Unique
+            $Version = $AllVersions | Sort-Object { [System.Version]$_ } -Descending | Select-Object -First 1
+            Write-Verbose -Message "$($MyInvocation.MyCommand): Latest version: $Version."
         }
-        Catch {
-            Throw "$($MyInvocation.MyCommand): Failed to extract version numbers from $uri"
+        catch {
+            throw "$($MyInvocation.MyCommand): Failed to extract version numbers from $uri. $($_.Exception.Message)"
         }
-
-        # Grab latest version number
-        $Version = ($AllVersions | Sort-Object { [System.Version]$_ } -Descending) | Select-Object -First 1
-        Write-Verbose -Message "$($MyInvocation.MyCommand): Latest version: $Version."
 
         # Grab latest Windows files
-        $LatestReleases = $FileNames -match $Version
+        $LatestReleases = ($FileNames -match $Version)[-1]
 
         # We need to rebase the timestamps from unix time, so need the Unix Epoch
         $UnixEpoch = ([System.DateTime] '1970-01-01Z').ToUniversalTime()
 
         # Build the output object for each release
-        ForEach ($Release in $LatestReleases) {
+        foreach ($Release in $LatestReleases) {
+
             # Construct the output; Return the custom object to the pipeline
             $PSObject = [PSCustomObject] @{
                 Version      = $Version
-                Architecture = Get-Architecture $Release
-                Date         = $UnixEpoch.AddSeconds($updateFeed.$Release.mtime)
-                Size         = $updateFeed.$Release.size
-                MD5          = $updateFeed.$Release.md5
-                Sha256       = $updateFeed.$Release.sha256
+                Architecture = Get-Architecture -String $Release
+                Date         = $UnixEpoch.AddSeconds($UpdateFeed.$Release.mtime)
+                Size         = $UpdateFeed.$Release.size
+                MD5          = $UpdateFeed.$Release.md5
+                Sha256       = $UpdateFeed.$Release.sha256
                 URI          = $res.Get.Update.Uri -replace "#replace", $release
             }
             Write-Output -InputObject $PSObject
         }
-    }
-    Else {
-        Write-Warning -Message "$($MyInvocation.MyCommand): unable to retrieve content from $Uri."
     }
 }