+++ Evergreen/Apps/Get-Microsoft.NET.ps1
@@ -1,4 +1,4 @@
-Function Get-Microsoft.NET {
+function Get-Microsoft.NET {
     <#
         .SYNOPSIS
             Returns the available Microsoft .NET Desktop Runtime versions and download URIs.
@@ -10,18 +10,18 @@ Function Get-Microsoft.NET {
     [OutputType([System.Management.Automation.PSObject])]
     [CmdletBinding(SupportsShouldProcess = $False)]
     param (
-        [Parameter(Mandatory = $False, Position = 0)]
+        [Parameter(Mandatory = $false, Position = 0)]
         [ValidateNotNull()]
         [System.Management.Automation.PSObject]
         $res = (Get-FunctionResource -AppName ("$($MyInvocation.MyCommand)".Split("-"))[1])
     )
 
     # Read the version number from the version URI
-    ForEach ($Channel in $res.Get.Update.Channels) {
+    foreach ($Channel in $res.Get.Update.Channels) {
 
         # Determine the version for each channel
         $Content = Invoke-RestMethodWrapper -Uri $($res.Get.Update.Uri -replace $res.Get.Update.ReplaceText, $Channel)
-        If ($Null -ne $Content) {
+        if ($null -ne $Content) {
 
             # Read last line of the returned content to retrieve the version number
             Write-Verbose -Message "$($MyInvocation.MyCommand): Returned: $Content."
@@ -39,12 +39,12 @@ Function Get-Microsoft.NET {
             Write-Verbose -Message "$($MyInvocation.MyCommand): Found $($Releases.releases.Count) release/s."
 
             # Step through each release type
-            ForEach ($Installer in $res.Get.Download.Installers) {
+            foreach ($Installer in $res.Get.Download.Installers) {
                 Write-Verbose -Message "$($MyInvocation.MyCommand): Build for type: $Installer."
                 Write-Verbose -Message "$($MyInvocation.MyCommand): Found $($Releases.releases[0].$Installer.files.count) files."
 
                 # Each installer includes multiple file types and platforms
-                ForEach ($File in $Releases.releases[0].$Installer.files) {
+                foreach ($File in $Releases.releases[0].$Installer.files) {
 
                     # Filter for .exe only so that we get Windows installers
                     $File | Where-Object { $_.name -match "\.exe$" } | ForEach-Object {
@@ -52,9 +52,10 @@ Function Get-Microsoft.NET {
                         # Build the output object; Output object to the pipeline
                         $PSObject = [PSCustomObject] @{
                             Version      = $Releases.releases[0].$Installer.version
-                            Architecture = Get-Architecture -String $_.rid
+                            Architecture = if ($_.rid.length -gt 0) { Get-Architecture -String $_.rid } else { Get-Architecture -String $_.url }
                             Installer    = $Installer
                             Channel      = $Channel
+                            Hash         = $_.hash
                             Type         = [System.IO.Path]::GetExtension($_.url).Split(".")[-1]
                             URI          = $_.url
                         }